+++ src/demo/constants/demo/index.js
@@ -1,10 +1,29 @@
+import React from 'react';
 import { isRequired, minLength } from '../../utils/validation';
+<<<<<<< HEAD
 import normalizeNumbers from '../../utils/normalization/number';
 import { type DataType } from '../../../lib/forms/fieldset';
 
 const DROPDOWN = [{ value: 'first', label: 'First' }, { value: 'second', label: 'Second' }];
 
 const FIELD_SETS: DataType = [
+=======
+import Check from '../../assets/icons/check.svg';
+
+const DROPDOWN = [{ value: 'first', label: 'First' }, { value: 'second', label: 'Second' }];
+
+const RATING = [
+  { value: 1, label: '1' },
+  { value: 2, label: '2' },
+  { value: 3, label: '3' },
+  { value: 4, label: '4' },
+  { value: 5, label: '5' },
+];
+
+const RATING_KEY = { left: 'Poor', right: 'Excellent' };
+
+const FIELD_SETS = [
+>>>>>>> master
   {
     fieldSet: [
       {
@@ -13,9 +32,9 @@ const FIELD_SETS: DataType = [
           {
             type: 'input',
             name: 'firstInput',
-            label: 'First Input',
+            label: 'Required Input',
+            required: true,
             validate: [isRequired, minLength(6)],
-            normalize: normalizeNumbers(16),
           },
           {
             type: 'dropdown',
@@ -23,6 +42,13 @@ const FIELD_SETS: DataType = [
             label: 'Dropdown Test',
             values: DROPDOWN,
             validate: isRequired,
+          },
+          {
+            type: 'input',
+            name: 'firstInput',
+            label: 'Required Input',
+            required: true,
+            validate: [isRequired, minLength(6)],
           },
           {
             type: 'dropdown',
@@ -34,7 +60,7 @@ const FIELD_SETS: DataType = [
           },
           {
             type: 'input',
-            name: 'secondInput',
+            name: 'pwInput',
             label: 'Password Input',
             inputType: 'password',
           },
@@ -43,6 +69,8 @@ const FIELD_SETS: DataType = [
             name: 'secondInput',
             label: 'Number Input',
             inputType: 'number',
+            min: 0,
+            max: 10,
           },
         ],
       },
@@ -54,6 +82,75 @@ const FIELD_SETS: DataType = [
             name: 'datePicker',
             label: 'Date Picker',
             validate: [isRequired],
+            showYearDropdown: true,
+            showMonthDropdown: true,
+            isOutsideRange: () => false,
+          },
+        ],
+      },
+      {
+        columns: { small: 1, large: 2 },
+        fields: [
+          {
+            type: 'daterangepicker',
+            name: 'dateRangePicker',
+            label: 'Date Range Picker',
+            validate: [isRequired],
+            showYearDropdown: true,
+            showMonthDropdown: true,
+          },
+        ],
+      },
+      {
+        columns: { small: 2 },
+        fields: [
+          {
+            type: 'radioButton',
+            name: 'accountType',
+            label: 'Individual',
+            value: 'individual',
+            validate: isRequired,
+          },
+          {
+            type: 'radioButton',
+            name: 'accountType',
+            label: 'Business',
+            value: 'company',
+            validate: isRequired,
+          },
+        ],
+      },
+      {
+        columns: { small: 1, large: 2 },
+        fields: [
+          {
+            type: 'checkbox',
+            name: 'checkbox',
+            label: 'Checkbox',
+          },
+        ],
+      },
+      {
+        columns: { small: 1, large: 2 },
+        fields: [
+          {
+            label: 'Toggle',
+            type: 'toggleButton',
+            name: 'toggle',
+            values: [{ label: 'label1', value: 1 }, { label: 'label2', value: 2 }],
+            validate: isRequired,
+          },
+        ],
+      },
+      {
+        columns: { small: 1, large: 2 },
+        fields: [
+          {
+            type: 'rating',
+            name: 'rating',
+            values: RATING,
+            icon: <Check />,
+            ratingKey: RATING_KEY,
           },
           {
             type: 'textarea',