+++ src/lib/forms/datepicker/index.js
@@ -1,52 +1,52 @@
 import React from 'react';
-import ReactDatePicker from 'react-datepicker';
-// $FlowFixMe
-import 'react-datepicker/dist/react-datepicker.css';
-import { Field, type Validator } from 'redux-form';
+
+import 'react-dates/initialize';
+import 'react-dates/lib/css/_datepicker.css';
+import { SingleDatePicker } from 'react-dates';
+
+import { Field } from 'redux-form';
 import styled from 'styled-components';
 import moment from 'moment';
+import Label from '../label';
 
 import Error from '../error';
-import { CLARK_SECONDARY } from '../../styles/colors';
-import { SPACING_SMALL } from '../../styles/spacing';
 
-type RenderPropsType = {
-  input: {
-    onBlur(x: string): void,
-    onFocus(): void,
-    onChange(): void,
-    value: string,
-  },
-  meta: {
-    touched: boolean,
-    error: ?string,
-  },
-};
+import './react-dates-overrides.css';
 
-export const renderDatePicker = ({ input, meta: { touched, error } }: RenderPropsType) => (
-  <div onBlur={() => input.onBlur(input.value)}>
-    <ReactDatePicker
-      selected={input.value ? moment(input.value) : null}
-      onFocus={input.onFocus}
-      onChange={input.onChange}
-      dateForm="MM/DD/YYYY"
-      placeholderText="MM/DD/YYYY"
+export const SingleDatePickerField = ({
+  meta: { active, error, touched },
+  input: { value = null, onChange, onFocus, onBlur },
+  placeholder = 'Select a date',
+  isOutsideRange,
+  numberOfMonths = 1,
+}) => (
+  <div>
+    <SingleDatePicker
+      date={value ? moment(value) : null}
+      onDateChange={e => (e ? onChange(e.toISOString()) : onChange(null))}
+      focused={active}
+      onFocusChange={({ focused }) => (focused ? onFocus() : onBlur())}
+      id="date"
+      placeholder={placeholder}
+      showDefaultInputIcon
+      hideKeyboardShortcutsPanel
+      isOutsideRange={isOutsideRange}
+      numberOfMonths={numberOfMonths}
     />
     <Error touched={touched} error={error} />
   </div>
 );
 
-type PropsType = {
-  name: string,
-  label: string,
-  columns: { small: number, large: number },
-  validate: Validator | Validator[],
-};
-
-const DatePicker = ({ name, label, columns, validate }: PropsType) => (
+const DatePicker = ({ name, label, columns, validate, required, ...props }) => (
   <Container>
-    <Label htmlFor={name}>{label}</Label>
-    <Field name={name} component={renderDatePicker} columns={columns} validate={validate} />
+    <Label name={name} label={label} required={required} />
+    <Field
+      name={name}
+      component={SingleDatePickerField}
+      columns={columns}
+      validate={validate}
+      {...props}
+    />
   </Container>
 );
 
@@ -54,10 +54,4 @@ export default DatePicker;
 
 const Container = styled.div`
   display: block;
-`;
-
-const Label = styled.label`
-  color: ${CLARK_SECONDARY};
-  display: block;
-  margin-bottom: ${SPACING_SMALL};
 `;