+++ src/lib/forms/radio-button/index.js
@@ -2,6 +2,7 @@ import React, { type Element } from 'react';
 import { Field } from 'redux-form';
 import styled from 'styled-components';
 
+import RequiredAsterisk from '../required-asterisk';
 import media from '../../styles/media-queries';
 import { SPACING_SMALL } from '../../styles/spacing';
 import { TYPE_SCALE_F5 } from '../../styles/type-scale';
@@ -9,7 +10,7 @@ import { CLARK_TERTIARY, WHITE } from '../../styles/colors';
 import { BORDER_RADIUS_F3 } from '../../styles/border-radius';
 import { BORDER_WIDTH_2 } from '../../styles/borders';
 
-type Props = {
+type PropsType = {
   name: string,
   label: string,
   index: number,
@@ -18,7 +19,7 @@ type Props = {
   icon: Element<any>,
 };
 
-const RadioButton = ({ name, label, index, value, placeholder, icon }: Props) => (
+const RadioButton = ({ name, label, index, value, placeholder, icon, required }: PropsType) => (
   <RadioButtonContainer>
     <Radio
       name={name}
@@ -29,7 +30,10 @@ const RadioButton = ({ name, label, index, value, placeholder, icon }: Props) =>
       index={index}
     />
     <Icon>{icon}</Icon>
-    <Label htmlFor={name}>{label}</Label>
+    <Label htmlFor={name}>
+      {label}
+      {required && <RequiredAsterisk />}
+    </Label>
   </RadioButtonContainer>
 );
 
@@ -62,7 +66,6 @@ const RadioButtonContainer = styled.div`
   height: 105px;
   position: relative;
   text-align: center;
-  z-index: 999;
 
   ${media.small`
     height: 150px;
@@ -73,7 +76,6 @@ const RadioButtonContainer = styled.div`
   */
   > ${Radio}:checked {
     background-color: ${CLARK_TERTIARY};
-    z-index: 0;
   }
 
   > ${Radio}:checked ~ label {