+++ src/lib/forms/textarea/index.js
@@ -3,6 +3,7 @@ import { Field } from 'redux-form';
 import styled from 'styled-components';
 
 import Error from '../error';
+
 import {
   SPACING_EXTRA_SMALL,
   SPACING_SMALL,
@@ -16,6 +17,7 @@ import { FONT_WEIGHT_100 } from '../../styles/font-weight';
 import { BORDER_WIDTH_1, renderBorderRadius } from '../../styles/borders';
 import { LINE_HEIGHT_COPY } from '../../styles/line-height';
 import { WHITE, CLARK_PRIMARY, CLARK_SECONDARY, GREY_25, ERROR_PRIMARY } from '../../styles/colors';
+import Label from '../label';
 
 const renderField = ({ input, index, placeholder, columns, type, meta: { touched, error } }) => (
   <div>
@@ -33,18 +35,12 @@ const renderField = ({ input, index, placeholder, columns, type, meta: { touched
 
 const Textarea = (props: { name: string, label: string }) => (
   <div>
-    <Label htmlFor={props.name}>{props.label}</Label>
+    <Label name={props.name} label={props.label} required={props.required} />
     <Field component={renderField} {...props} />
   </div>
 );
 
 export default Textarea;
-
-const Label = styled.label`
-  color: ${CLARK_SECONDARY};
-  display: block;
-  margin-bottom: ${SPACING_SMALL};
-`;
 
 const FormTextarea = styled.textarea`
   ${TYPE_SCALE_F4};
@@ -64,7 +60,6 @@ const FormTextarea = styled.textarea`
   margin-bottom: ${props => (props.showError ? 0 : SPACING_SMALL)};
   background-color: ${props => (props.showError ? WHITE : ERROR_PRIMARY)};
   transition: 0.5s;
-  z-index: ${props => (props.showError ? 0 : 9)};
   ${props => renderBorderRadius(props)};
 
   &::placeholder {