+++ src/lib/index.js
@@ -1,5 +1,25 @@
 import Fieldset from './forms/fieldset';
 import ProgressHeader from './forms/progress-header';
+import Checkbox from './forms/checkbox';
+import DatePicker from './forms/datepicker';
+import DateRangePicker from './forms/date-range-picker';
+import Dropdown from './forms/dropdown';
+import Error from './forms/error';
+import Input from './forms/input';
+import RadioButton from './forms/radio-button';
+import Textarea from './forms/textarea';
+import Rating from './forms/rating';
 
 export default Fieldset;
-export { ProgressHeader };
+export {
+  ProgressHeader,
+  Checkbox,
+  DatePicker,
+  DateRangePicker,
+  Dropdown,
+  Error,
+  Input,
+  RadioButton,
+  Textarea,
+  Rating,
+};