+++ CHANGELOG.md
@@ -1,6 +1,52 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## 0.8.8 - 2018-06-07
+### Fixed
+- checkbox component wasn't hooked up correctly and now it is
+
+## 0.8.7 - 2018-05-14
+### Changed
+- made all dependencies dev dependencies because we build before publishing
+
+
+## 0.8.6 - 2018-05-14
+### Changed
+- Remove z-indexing from form components to fix myriad of visual issues
+
+## 0.8.2 - 2018-05-02
+### Changed
+- No logo in progress bar
+
+## 0.8.1 - 2018-05-01
+### Changed
+- Ratings component styles
+
+## 0.8.0 - 2018-05-01
+### Added
+- Ratings component
+
+## 0.7.3 - 2018-04-18
+### Refactored
+- ToggleButtons
+- Spreading props to input
+
+## 0.7.2 - 2018-04-13
+### Added
+- ToggleButtons
+
+## 0.7.1 - 2018-04-11
+### Changed
+- Styled date picker and added date range picker.
+
+## 0.7.0 - 2018-03-21
+### Changed
+- Added all form components to exports.
+
+## 0.6.1 - 2018-03-19
+### Changed
+- Allow datepicker to accept props for `react-datepicker`
+
 ## 0.6.0 - 2018-03-06
 ### Changed
 - Readme updated the Updating library section